recipe-backend/requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
autopep8==1.6.0
boto3==1.20.40
botocore==1.23.40
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.10
click==8.1.3
databases==0.7.0
exceptiongroup==1.1.1
fastapi==0.97.0
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.19.1
httptools==0.5.0
idna==2.10
jmespath==0.10.0
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.22.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.5
protobuf==3.17.3
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.10.9
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
toml==0.10.2
typing_extensions==4.6.3
uritemplate==3.0.1
urllib3==1.26.6
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
